services/event_ingestor/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
django-environ>=0.11,<1.0
gunicorn>=21.2,<23.0
psycopg[binary]>=3.1,<4.0
